requirements.txt
Django==4.2.8
python-decouple==3.8
django-debug-toolbar==4.2.0